Make suggested changes; flatten module hierarchy.
