+++ Cargo.toml
@@ -11,18 +11,14 @@ exclude = ["src/bin/", "examples/", "scripts/"]
 license = "MIT OR Apache-2.0"
 readme = "README.md"
 
-[features]
-default = ["cmplx"]
-cmplx = ["dep:num-complex"]
-
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [package.metadata.docs.rs]
-rustdoc-args = [ "--html-in-header", "katex-header.html", "--cfg", "docsrs"]
+rustdoc-args = ["--html-in-header", "katex-header.html", "--cfg", "docsrs"]
 
 [dependencies]
 lambert_w = { version = "1.0.0", default-features = false, features = ["std"] }
-num-complex = { version = "0.4", optional = true }
+num-complex = { version = "0.4" }
 
 [dev-dependencies]
 peroxide = { version = "0.39", features = ["plot"] }
